requirements.txt
robotframework>=4.1.0
robotframework-seleniumlibrary>=5.1.3
robotframework-requests>=0.9.2
pylint>=2.12.0
